Unify Tryout shell into TryoutShell component
