--- apps/www/app/[locale]/(app)/(shared)/try-out/(catalog)/layout.tsx
@@ -1,6 +1,0 @@
-import { AppShell } from "@/components/sidebar/app-shell";
-
-/** Renders the shared plain shell for the tryout hub and product catalog routes. */
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <AppShell>{children}</AppShell>;
-}
