--- apps/www/app/[locale]/(app)/(shared)/try-out/(session)/[product]/[slug]/loading.tsx
@@ -1,4 +1,0 @@
-/** Renders the session loading state inside the shared tryout shell. */
-export default function Loading() {
-  return <div className="min-h-svh" />;
-}
