+++ apps/www/app/[locale]/(app)/(shared)/try-out/loading.tsx
@@ -1,4 +1,4 @@
-/** Renders the catalog loading state inside the shared tryout shell. */
+/** Renders the tryout route fallback inside the shared tryout shell. */
 export default function Loading() {
   return <div className="min-h-svh" />;
 }
